Silent pushes *do not* use websocket in order to call completion handler
Instead they use the HTTP api, put in place for our censorship
circumvention.

The reason we avoid using websockets, is that there's no real way to
know that we're "done" fetching messages, since it's server-push based.

It's important to know that we're done so we can call the notification
completion handler. The completion handler lets the OS know that we're
done, and can be safely suspended. Otherwise we risk overspending our
background time budget, reducing the chance that we're allowed to be
launched from the background in the future.

// FREEBIE